Extract sidecar load-path store
