+++ docs/dev/plans/refactor-4-pass2-storage-sidecars.md
@@ -1,14 +1,14 @@
 # Refactor-4 Pass 2 - Storage and Sidecar Owner Proposal
 
 **Date:** 2026-04-25.
-**Worktree:** `Parsek-refactor-4-pass2-sidecar-save`, branch
-`refactor-4-pass2-sidecar-save`.
-**Base:** `origin/main` after PR #552 (`refactor-4-pass2`) merged; this
-slice was originally cut from PR #552's branch.
+**Worktree:** `Parsek-refactor-4-pass2-sidecar-load`, branch
+`refactor-4-pass2-sidecar-load`.
+**Base:** stacked on PR #554's `refactor-4-pass2-sidecar-save` branch; PR #554
+itself is intentionally left unmerged for the later 0.9.1 batch.
 **Status:** Proposal plus implementation checkpoints. The
-`SidecarFileCommitBatch` and save-path `RecordingSidecarStore` slices are
-complete; load-path, codec, and tree-record moves remain proposal-only until a
-separate approval.
+`SidecarFileCommitBatch`, save-path `RecordingSidecarStore`, and load-path
+`RecordingSidecarStore` slices are complete; codec and tree-record moves remain
+proposal-only until a separate approval.
 
 ## Guardrails
 
@@ -333,3 +333,38 @@ non-injection gate passed 8,707 tests, and `InjectAllRecordings` passed 3 tests.
 
 Runtime canaries are not required for this save-only mechanical move, but the
 load-path extraction will need the runtime sidecar probe canary before merge.
+
+## Implementation Checkpoint - RecordingSidecarStore Load Path
+
+Approved third slice stacked on PR #554's save-path branch and kept to the load
+path only:
+
+- Moved `LoadRecordingFiles`, `LoadRecordingFilesFromPathsForTesting`, sidecar
+  load-failure marking/clearing, sidecar epoch validation, snapshot sidecar
+  load summary/fallback policy, post-hydration loop repairs, terminal-orbit
+  backfill, and endpoint backfill into `RecordingSidecarStore`.
+- Kept `RecordingStore` wrappers, `SnapshotSidecarLoadState`, and
+  `SnapshotSidecarLoadSummary` stable for existing production and test call
+  sites.
+- Preserved the trajectory load order exactly: probe, supported/id/epoch gates,
+  deserialize, loop migration/degenerate-loop repair, terminal-orbit backfill,
+  endpoint backfill, snapshot sidecars, failure flagging.
+- Left trajectory and snapshot codec dispatch in the existing codec owners.
+- Added direct xUnit coverage for
+  `RecordingSidecarStore.LoadRecordingFilesFromPathsForTesting`, including the
+  snapshot-failure path that must preserve hydrated trajectory points.
+
+Validation completed:
+
+```powershell
+dotnet test Source/Parsek.Tests/Parsek.Tests.csproj --filter "FullyQualifiedName~RecordingStorageRoundTripTests|FullyQualifiedName~SnapshotSidecarCodecTests|FullyQualifiedName~TrajectorySidecarBinaryTests|FullyQualifiedName~Bug270SidecarEpochTests|FullyQualifiedName~FormatVersionTests|FullyQualifiedName~TrackSectionSerializationTests|FullyQualifiedName~LoopIntervalLoadNormalizationTests|FullyQualifiedName~QuickloadResumeTests"
+dotnet test Source/Parsek.Tests/Parsek.Tests.csproj --filter FullyQualifiedName!~InjectAllRecordings
+```
+
+Latest run: focused storage slice passed 236 tests and the non-injection gate
+passed 8,708 tests. `InjectAllRecordings` is currently blocked because
+`KSP.log` and `GameData/Parsek/Plugins/Parsek.dll` are locked by a running KSP
+process.
+
+The runtime sidecar probe canary and a manual save/load or quickload canary are
+still required before merging this load-path slice.
